Split code + Fix Jest testing
