+++ index.spec.js
@@ -1,45 +1,8 @@
-const TwitterBot = require('./index.js');
-const Twit = require('twit');
+const Index = require('./index.js')
 
-test('Only a mock test', () => {
-    expect(true).toBe(true);
-});
-
-
-// Smoke Test
-describe('Smoke Tests', () => {
-    describe('Bot', () => {
-        it('should exists as object', () => {
-            expect(typeof TwitterBot.Bot).toBe('object');
-        });
-
-        it('should be instance of Twit', () => {
-            expect(TwitterBot.Bot).toBeInstanceOf(Twit);
-        });
-
-        it('should Bot config have twitter access properties', () => {
-            const expected = {
-                consumer_key: expect.anything(),
-                consumer_secret: expect.anything(),
-                access_token: expect.anything(),
-                access_token_secret: expect.anything(),
-            };
-            expect(TwitterBot.Bot.config).toEqual(expect.objectContaining(expected))
-        });
-
-    });
-
-    describe('BotRetweet', () => {
-        it('should BotRetweet exist and be a function', () => {
-            expect(typeof TwitterBot.BotRetweet).toBe('function');
-        });
-    });
-
-    describe('isReply', () => {
-        it('should exist and be a function', () => {
-            expect(typeof TwitterBot.isReply).toBe('function');
-        });
-    });
-
+describe('Index', () => {
+    const BotInitMock = jest.fn(Index.BotInit)
+    it('should BotInit to be defined', () => {
+        expect(BotInitMock).toBeDefined();
+    })
 })
-
